refactor JavaProcess/ProcessInfo
